feat(repository) : 리포지토리 인터페이스 생성
