+++ src/main/java/roomescape/service/ReservationService.java
@@ -5,31 +5,31 @@ import org.springframework.stereotype.Service;
 import roomescape.dto.ReservationRequestDto;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
-import roomescape.repository.ReservationRepository;
+import roomescape.repository.JdbcReservationRepository;
 
 @Service
 public class ReservationService {
-    private final ReservationRepository reservationRepository;
+    private final JdbcReservationRepository jdbcReservationRepository;
     private final ReservationTimeService reservationTimeService;
 
-    public ReservationService(ReservationRepository reservationRepository,
+    public ReservationService(JdbcReservationRepository jdbcReservationRepository,
                               ReservationTimeService reservationTimeService) {
-        this.reservationRepository = reservationRepository;
+        this.jdbcReservationRepository = jdbcReservationRepository;
         this.reservationTimeService = reservationTimeService;
     }
 
     public List<Reservation> getAllReservations() {
-        return reservationRepository.getAllReservations();
+        return jdbcReservationRepository.getAllReservations();
     }
 
     public Reservation addReservation(ReservationRequestDto reservationRequestDto) {
         ReservationTime reservationTime = reservationTimeService.getReservationTimeById(
                 reservationRequestDto.time_id());
-        return reservationRepository.addReservation(reservationRequestDto, reservationTime);
+        return jdbcReservationRepository.addReservation(reservationRequestDto, reservationTime);
     }
 
     public Integer deleteReservation(long id) {
-        return reservationRepository.deleteReservation(id);
+        return jdbcReservationRepository.deleteReservation(id);
     }
 
 }
